rename: subCategory -> genre 이름 변경 및 폴더 구조 변경
