+++ README.md
@@ -1,1 +1,3 @@
 # fe-kakaopage
+
+- [ ] 웹툰 말고도 웹소설, 영화 Category 도 사용한다고 했을 때 어떻게 Render 할 지 ?
